wip: #WB2-1603, add delete modal
