+++ backend/src/main/resources/jsonschema/page.json
@@ -3,12 +3,27 @@
   "type": "object",
   "additionalProperties": false,
   "properties": {
-    "title": { "type": "string" },
-    "content": { "type": "string" },
-    "parentId": { "type":  ["string", "null"] },
-    "isIndex": { "type": "boolean"},
-    "wasIndex": { "type": "boolean" },
-    "isVisible": { "type": "boolean" }
+    "title": {
+      "type": "string"
+    },
+    "content": {
+      "type": "string"
+    },
+    "parentId": {
+      "type": [
+        "string",
+        "null"
+      ]
+    },
+    "isIndex": {
+      "type": "boolean"
+    },
+    "wasIndex": {
+      "type": "boolean"
+    },
+    "isVisible": {
+      "type": "boolean"
+    }
   },
   "required": [
     "title"
